requirements.txt
# transformers==4.35.2
# torch==2.1.1
# deep-translator==1.11.4
# streamlit==1.28.1
# googletrans==4.0.0
# Pillow==10.1.0

# Core Dependencies
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.14.0
sentencepiece>=0.1.99

# Translation & NLP
deep-translator>=1.11.0
googletrans>=4.0.0

# Web Framework
streamlit>=1.28.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0

# Utilities
Pillow>=10.0.0
tqdm>=4.66.0
requests>=2.31.0

# Development (optional)
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
